refactor: Move all methods handling data to util.DataUtil
