+++ src/winter/util/DataUtil.java
@@ -1,14 +1,19 @@
 package winter.util;
 
 import java.lang.reflect.AnnotatedElement;
+import java.time.Instant;
+import java.time.ZoneOffset;
+import java.time.format.DateTimeFormatter;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import jakarta.servlet.http.HttpServletRequest;
 import winter.data.annotation.http.RequestParam;
 import winter.data.enumdata.RequestParamType;
 import winter.data.exception.client.InvalidFormDataException;
 
 public class DataUtil extends Utility {
+    /* ----------------------------- Parsing methods ---------------------------- */
     public static Object parseObject(Class<?> objType, String value) {
         if (value == null) {
             return null;
@@ -83,5 +88,37 @@ public class DataUtil extends Utility {
         } catch (NumberFormatException e) {
             return false;
         }
+    }
+
+    public static boolean isValidPackageName(String packageName) {
+        String packageNameRegex = "^(\\w+)(\\.(\\w+))*$";
+        Pattern pattern = Pattern.compile(packageNameRegex);
+
+        if (packageName == null || packageName.isEmpty()) {
+            return false;
+        }
+        return pattern.matcher(packageName).matches();
+    }
+
+    /* --------------------------- Generation methods --------------------------- */
+    public static String generateTimestampFilename(String originalFilename) {
+        Instant now = Instant.now();
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyyMMdd_HHmmssSSS").withZone(ZoneOffset.UTC);
+        String timestamp = formatter.format(now);
+
+        if (originalFilename != null && !originalFilename.isEmpty()) {
+            int lastDotIndex = originalFilename.lastIndexOf(".");
+            if (lastDotIndex != -1) {
+                String extension = originalFilename.substring(lastDotIndex);
+                return "file_" + timestamp + extension;
+            }
+        }
+
+        return "file_" + timestamp + ".bin"; // Default to .bin if no extension
+    }
+
+    /* --------------------------- Extraction methods --------------------------- */
+    public static String extractURIMapping(HttpServletRequest req) {
+        return req.getRequestURI().substring(req.getContextPath().length());
     }
 }
