--- src/winter/util/UrlUtil.java
@@ -1,14 +1,0 @@
-package winter.util;
-
-import jakarta.servlet.http.HttpServletRequest;
-
-public class UrlUtil extends Utility {
-    public static String extractTargetURL(HttpServletRequest req) {
-        return req.getRequestURI().substring(req.getContextPath().length());
-    }
-
-    public static String extractLastURL(HttpServletRequest req) {
-        String[] splitRequest = req.getRequestURL().toString().split("/");
-        return splitRequest[splitRequest.length - 1];
-    }
-}
